rename sf fish tts
